made cheque and modify PO

+++ src/microapps/priceChartUploadApp/PriceChart.jsx
@@ -50,22 +50,26 @@ export default function PriceChart() {
         // const wbName = wb.SheetNames[0]
         // const ws = wb.Sheets[wbName]
         // const xlData = xlxs.utils.sheet_to_json(ws)
-        const xlData = getxlData()
-        console.log(xlData);
+        if (xlfile) {
+            const xlData = getxlData()
+            console.log(xlData);
 
-        try {
-            ApiService.setHeader();
+            try {
+                ApiService.setHeader();
 
-            const response = await ApiService.delete('/priceChartUpload/procedure')
-            if (response.data.isSuccess) {
-                const res = await ApiService.post('/priceChartUpload/procedure', xlData)
-                if (res.data.isSuccess) {
-                    console.log(res.data.documents);
-                    navigate(`/${rootPath}/pricechartlist`);
+                const response = await ApiService.delete('/priceChartUpload/procedure')
+                if (response.data.isSuccess) {
+                    const res = await ApiService.post('/priceChartUpload/procedure', xlData)
+                    if (res.data.isSuccess) {
+                        console.log(res.data.documents);
+                        navigate(`/${rootPath}/pricechartlist`);
+                    }
                 }
+            } catch (e) {
+                errorMessage(e, null)
             }
-        } catch (e) {
-            errorMessage(e, null)
+        } else {
+            infoNotification("Please select any file to upload")
         }
     }
 
@@ -101,7 +105,7 @@ export default function PriceChart() {
             <Form onSubmit={handleSubmit(onSubmit)} className="pct-app-content">
                 <Container className="pct-app-content-header m-0 mt-2 pb-2" style={{ borderBottom: '1px solid black' }} fluid>
                     <Row>
-                        <Col><h3>Price Chart Upload</h3></Col>
+                        <Col><h3>PRICE CHART UPLOAD</h3></Col>
                     </Row>
                     <Row>
                         <Col>
@@ -117,7 +121,7 @@ export default function PriceChart() {
                     <Container className="mt-2" fluid>
                         <Row>
                             <Form.Group as={Col} md="12" className="mb-2">
-                                <Form.Label className="m-0">Upload file</Form.Label>
+                                <Form.Label className="m-0">UPLOAD FILE</Form.Label>
                                 <Form.Control
                                     type="file"
                                     id="file"
